Fixed conversion warnings on framework tests (#2606)
---------

+++ Fw/Port/OutputSerializePort.cpp
@@ -22,7 +22,7 @@ namespace Fw {
     void OutputSerializePort::toString(char* buffer, NATIVE_INT_TYPE size) {
 #if FW_OBJECT_NAMES == 1
         FW_ASSERT(size > 0);
-        if (snprintf(buffer, size, "Output Serial Port: %s %s->(%s)", this->m_objName.toChar(), this->isConnected() ? "C" : "NC",
+        if (snprintf(buffer, static_cast<size_t>(size), "Output Serial Port: %s %s->(%s)", this->m_objName.toChar(), this->isConnected() ? "C" : "NC",
                      this->isConnected() ? this->m_connObj->getObjName() : "None") < 0) {
             buffer[0] = 0;
         }
